Added documentation in README

+++ src/ESP32_Supabase.h
@@ -52,7 +52,7 @@ public:
   int insert(String table, String json, bool upsert);
   Supabase &select(String colls);
   Supabase &update(String table);
-  int upload(String bucket, String filename, String mime_type, Stream *stream, uint64_t size);
+  int upload(String bucket, String filename, String mime_type, Stream *stream, uint32_t size);
   int upload(String bucket, String filename, String mime_type, uint8_t *buffer, uint32_t size);
 
   // Comparison Operator
